requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
chatterbox-tts
pyyaml>=6.0
numpy>=1.24.0